Rename core modules to be more semantic
