+++ src/content-scripts/handlers/chromeStorageHandlers.ts
@@ -4,13 +4,20 @@ import store from '../store';
 import { AppModes } from 'content-scripts/store/appModes/types';
 import chromeStorage from 'common/chrome-storage';
 
+export default {
+  appModes: {
+    init,
+    deinit,
+  },
+};
+
 const storageKeysToAppMode = {
   [chromeKeys.DISABLED_EXTENSION]: 'isExtensionDisabled',
   [chromeKeys.ANNOTATION_MODE_PAGES]: 'annotationModePages',
   [chromeKeys.DISABLED_PAGES]: 'disabledPages',
 };
 
-export default function initializeChromeStorageHandlers() {
+function init() {
   chromeStorage.onChanged.addListener((changes, namespace) => {
     for (const key of Object.keys(changes)) {
       if (storageKeysToAppMode[key]) {
@@ -20,4 +27,8 @@ export default function initializeChromeStorageHandlers() {
       }
     }
   });
+}
+
+function deinit() {
+  // (todo)
 }
