+++ src/content-scripts/modules/index.ts
@@ -1,0 +1,3 @@
+export { default as highlightManager } from './highlightManager';
+export { default as annotationLocator } from './annotationLocator';
+export { default as appComponent } from './appComponent';
